tests/server/test_werkzeug.py: Fix test for github actions
Fix: TypeError: cannot unpack non-iterable TestResponse object

Add support for testing with werkzeug 2.1 while keeping tuple unpacking
for werkzeug 1.0. This should fix the github checks for pull requests.

The github checks for 3.9 now use Werkzeug 2.1 which no longer
provides a TestResponse as a tuple, resulting in this TypeError.

3.6 shows this warning: DeprecationWarning: The test client no longer
returns a tuple, it returns a 'TestResponse'. Tuple unpacking is
deprecated and will be removed in Werkzeug 2.1. Access the attributes
'data', 'status', and 'headers' instead.