aio-pika integration and backend updated for aio-pika 8.0
aio-pika 7.0 and 8.0 changed the API in a few places and have
fixes and reliabiltiy improvements over the older 6.8.

While the API is largely compatible, the special case of specifying
a results queue with certain queue properties using result_queue_name
and result_queue_args needs an update.

Update mypy type annotations for the new abc types of aio-pika 7.0.