aio_pika client: Fix handling aio_pika.Client(result_queue_name)
In xjsonrpc/client/backend/aio_pika.py there was an oversight where
when result_queue_name and result_queue_args) are used to specify
the use of a specific results queue, the connection, which the
queue gets from the channel is wrongly passed to the __init__ method.

This caused that all arguments were assigned to wrong method attrs
and the resulting queue could not be instantiated. Fix this oversight.

Also add am example demonstrating the use of such specific result queue.