Fix aio_pika client to never throw AssertionErrors
Issue #91 shows an AssertionError when a connected aio_pika client
which didn't need to open a result queue calls client.close().

It should be possible to call client.close() also in this case
to close the connection to the RabbitMQ server, because when
either the client was interrupted, had nothing to do before
closing, it, this is the normal state.

Also replace the other assertions in the aio_pika client's close()
with simple checks which don't abort the function to close any
other remaining connection elements or futures which are not
properly handled when simply aborting with an assertion.

For details and a test case, see issue #91.